requirements.txt
blessings==1.6
bluedot==1.0.4
coloredlogs==7.3
dbus-python==1.2.4
Events==0.3
humanfriendly==4.4.1
humanize==0.5.1
keyboard==0.11.0
monotonic==1.3
netaddr==0.7.19
numpy==1.13.1
observer==0.1.0
pbkdf2==1.3
pyserial==3.4
python-networkmanager==2.0.1
PyYAML==3.12
six==1.11.0
wcwidth==0.1.7
